ACO for VRP
Contains many hacks.

At the moment, ACO applies penalty for exceeding the vehicle capacity
only after the whole path have been constructed.

It might be better if the penalty was applied to the probabilies, so
that the ant will know that going to a vertex which demand he cannot
meet will have a high cost.